refactor(shardnode): reorganize shardnode cli tools
with: #1000685347
